Exception handling

+++ src/main/java/com/rogeriogregorio/bookmanagementsystem/services/BookService.java
@@ -2,11 +2,11 @@ package com.rogeriogregorio.bookmanagementsystem.services;
 
 import com.rogeriogregorio.bookmanagementsystem.dto.BookDTO;
 import com.rogeriogregorio.bookmanagementsystem.entities.BookEntity;
+import com.rogeriogregorio.bookmanagementsystem.exceptions.*;
 import com.rogeriogregorio.bookmanagementsystem.repositories.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 
@@ -19,6 +19,9 @@ public class BookService {
     @Transactional(readOnly = true)
     public List<BookDTO> findAllBooks() {
 
+        if (bookRepository.findAll().isEmpty()){
+            throw new BookNotFoundException("Nenhum livro encontrado.");
+        }
         return bookRepository
                 .findAll()
                 .stream()
@@ -29,33 +32,59 @@ public class BookService {
     @Transactional(readOnly = true)
     public BookDTO findBookById(Long id) {
 
-        return new BookDTO(bookRepository.findById(id).get());
+        return new BookDTO(bookRepository.findById(id).orElseThrow(
+                () -> new BookNotFoundException("Livro não encontrado com o ID: " + id + ".")));
     }
 
     @Transactional(readOnly = false)
     public List<BookDTO> createBook(BookEntity bookEntity) {
 
-        bookRepository.save(bookEntity);
+        if (bookRepository.existsById(bookEntity.getId())) {
+            throw new BookAlreadyExistsException("Livro já existe com o ID: " + bookEntity.getId() + ".");
+        }
+
+        try {
+            bookRepository.save(bookEntity);
+        } catch (Exception e) {
+            throw new BookCreationException("Erro ao criar o livro: " + e.getMessage() + ".");
+        }
+
         return findAllBooks();
     }
 
     @Transactional(readOnly = false)
     public List<BookDTO> updateBook(BookEntity bookEntity) {
 
-        bookRepository.save(bookEntity);
+        if (!bookRepository.existsById(bookEntity.getId())) {
+            throw new BookNotFoundException("Livro não encontrado com o ID: " + bookEntity.getId() + ".");
+        }
+
+        try {
+            bookRepository.save(bookEntity);
+        } catch (Exception e) {
+            throw new BookUpdateException("Erro ao atualizar o livro: " + e.getMessage() + ".");
+        }
+
         return findAllBooks();
     }
 
     @Transactional(readOnly = false)
     public List<BookDTO> deleteBook(Long id) {
 
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Livro não encontrado com o ID: " + id + ".");
+        }
+
         bookRepository.deleteById(id);
+
         return findAllBooks();
     }
 
     @Transactional(readOnly = true)
     public List<BookDTO> findBookByTitleOrAuthor(String titleOrAuthor) {
+
         String searchTerm = "%" + titleOrAuthor + "%";
+        
         return bookRepository.findByTitleOrAuthor(searchTerm)
                 .stream()
                 .map(BookDTO::new)
